third_party/blink/web_tests/platform/generic/fast/events/constructors/gamepad-event-constructor-expected.txt

This tests the constructor for the GamepadEvent DOM class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new GamepadEvent('eventType').bubbles is false
PASS new GamepadEvent('eventType').cancelable is false
PASS new GamepadEvent('eventType').gamepad is null
PASS new GamepadEvent('eventType', { bubbles: false }).bubbles is false
PASS new GamepadEvent('eventType', { bubbles: true }).bubbles is true
PASS new GamepadEvent('eventType', { cancelable: false }).cancelable is false
PASS new GamepadEvent('eventType', { cancelable: true }).cancelable is true
PASS new GamepadEvent('eventType', { gamepad: gamepad }).gamepad is gamepad
PASS successfullyParsed is true

TEST COMPLETE
